requirements.txt
# API and Web
requests>=2.28.1

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=10.0.0
pydantic>=1.10.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
alembic>=1.8.0

# Orchestration
prefect>=2.7.0


# Logging and Monitoring
python-json-logger>=2.0.4

# Configuration
python-dotenv>=0.21.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.1.0
isort>=5.12.0
flake8>=6.0.0
pre-commit>=3.0.0
